densify is slow, add SE
